feat(windows): native interop, PnP device management, WLAN, radio and probes
- wlanapi P/Invoke verified against SDK 26100 headers, with runtime struct layout self-check
- NativeWifiManager: SafeHandle, ACM/MSM notifications, async scan with per-adapter serialisation
  and timeout, profile list, connect/disconnect with state verification, BSS/available network merge
- Location privacy detection for ERROR_ACCESS_DENIED on WLAN APIs (never bypassed)
- SetupAPI/CfgMgr32 PnP inventory incl. NDIS driver key properties, physical device classification
  reused from Core, CM_Enable_DevNode/CM_Disable_DevNode plus on-demand elevated helper client
- IP Helper interface and route inventory, optional interface metric management (default off)
- Multi-signal connectivity probe: TCP/HTTP(S)/DNS/optional ICMP, per-interface source binding,
  captive portal interception detection, threshold based verdict
- Windows.Devices.Radios controller with native radio-state fallback
- Shell_NotifyIcon tray icon on a hidden Win32 window, HKCU run-at-logon registration

+++ src/NetworkGuardian.Core/Models/ProbeModels.cs
@@ -80,7 +80,7 @@ public sealed record DefaultRouteInfo
 
     public required uint InterfaceIndex { get; init; }
 
-    public uint? InterfaceLuid { get; init; }
+    public ulong? InterfaceLuid { get; init; }
 
     public required string NextHop { get; init; }
